backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
gunicorn
flask-jwt-extended
flask-bcrypt
google-generativeai
groq
python-dotenv
pydantic
pytest
flask-executor
pypdf
psycopg2-binary
email-validator
langchain
langchain-groq
langchain-community
chromadb
sentence-transformers
langchain-chroma
tiktoken
